# Conflicts:
#	.idea/deploymentTargetSelector.xml
#	app/build.gradle.kts
#	app/src/main/java/net/invictusmanagement/invictuskiosk/data/remote/RestClient.kt
#	app/src/main/java/net/invictusmanagement/invictuskiosk/data/repository/VideoCallRepositoryImpl.kt
#	app/src/main/java/net/invictusmanagement/invictuskiosk/di/AppModule.kt
#	app/src/main/java/net/invictusmanagement/invictuskiosk/presentation/MainViewModel.kt
#	app/src/main/java/net/invictusmanagement/invictuskiosk/presentation/coupon_detail/CouponDetailsScreen.kt
#	app/src/main/java/net/invictusmanagement/invictuskiosk/presentation/home/HomeScreen.kt
#	app/src/main/java/net/invictusmanagement/invictuskiosk/presentation/home/components/URLVideoPlayer.kt
#	app/src/main/java/net/invictusmanagement/invictuskiosk/presentation/leasing_office/LeasingOfficeScreen.kt
#	app/src/main/java/net/invictusmanagement/invictuskiosk/presentation/qr_code_scanner/QRScannerScreen.kt
#	app/src/main/java/net/invictusmanagement/invictuskiosk/presentation/qr_code_scanner/QRScannerViewModel.kt
#	app/src/main/java/net/invictusmanagement/invictuskiosk/presentation/residents/ResidentsScreen.kt
#	app/src/main/java/net/invictusmanagement/invictuskiosk/presentation/unlock/UnlockScreen.kt
#	app/src/main/java/net/invictusmanagement/invictuskiosk/presentation/vacancy/VacancyScreen.kt
#	app/src/main/java/net/invictusmanagement/invictuskiosk/presentation/video_call/VideoCallScreen.kt
#	app/src/main/java/net/invictusmanagement/invictuskiosk/presentation/video_call/VideoCallViewModel.kt

+++ app/src/main/java/net/invictusmanagement/invictuskiosk/domain/model/ServiceKey.kt
@@ -1,11 +1,11 @@
 package net.invictusmanagement.invictuskiosk.domain.model
 
 data class ServiceKey(
-    val accessLogId: Int=0,
-    val accessPointId: Int,
+    val accessLogId: Long = 0,
+    val accessPointId: Long,
     val isValid: Boolean = false,
     val key: String,
-    val keyId: Int = 0,
-    val serviceKeyUsageId: Int =0,
-    val snapShotImageId: Int=0
+    val keyId: Long = 0L,
+    val serviceKeyUsageId: Long = 0,
+    val snapShotImageId: Long = 0
 )
